feat (front): implementando a pesquisa de entradas e saidas por categoria
