+++ front/src/renderer/src/views/pages/Dashboard/components/LoaderContainer.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const LoaderContainer = styled.div`
+  height: 32.0rem;
+  width: 100%;
+`;
