+++ front/src/renderer/src/views/pages/Dashboard/sections/CategoriesSection/ExpensesSection/index.tsx
@@ -1,14 +1,15 @@
 import { Expense } from '@renderer/app/entities/Expense';
+import { useExpensesByTypeQuery } from '@renderer/app/hooks/queries/useExpensesByTypeQuery';
 import { formatCurrency } from '@renderer/app/utils/formatCurrency';
 import { DashboardCategoryIcon } from '@renderer/assets/Icons/dashboard/DashboardCategoryIcon';
-import { SelectedCategory } from '..';
+import { useCallback, useState } from 'react';
+import ExpensesModal from '../../../components/ExpensesModal';
 import { Item } from '../../../components/Item';
 import { SectionTitle } from '../styles';
 import { Container } from './styles';
 
 interface ExpenseSectionProps {
   expenses: Expense[];
-  onSelect?(category: SelectedCategory): void;
 }
 
 export const translateExpenseType = {
@@ -22,12 +23,42 @@ export const translateExpenseType = {
   garrisons: 'Guarnição',
 };
 
-export default function ExpensesSection({
-  expenses,
-  onSelect,
-}: ExpenseSectionProps) {
+export default function ExpensesSection({ expenses }: ExpenseSectionProps) {
+  const [isOpenExpenseModal, setIsOpenExpenseModal] = useState(false);
+  const [selectedExpenseType, setSelectedExpense] = useState<string>('');
+
+  const handleOpenExpenseModal = useCallback((expense: Expense) => {
+    setSelectedExpense(expense.type.toLowerCase());
+    setIsOpenExpenseModal(true);
+  }, []);
+
+  const handleCloseExpenseModal = useCallback(() => {
+    setSelectedExpense('');
+    setIsOpenExpenseModal(false);
+  }, []);
+
+  const { expenses: selectedTypeExpenses, isLoading } = useExpensesByTypeQuery({
+    type: selectedExpenseType,
+    enabled: isOpenExpenseModal,
+  });
+
   return (
     <Container>
+      {isOpenExpenseModal && (
+        <ExpensesModal
+          open
+          hasAction={false}
+          isLoading={isLoading}
+          expensesHistory={selectedTypeExpenses.history}
+          onClose={handleCloseExpenseModal}
+          title={
+            translateExpenseType[
+              selectedExpenseType as keyof typeof translateExpenseType
+            ]
+          }
+        />
+      )}
+
       <SectionTitle>Categorias de saídas</SectionTitle>
       <div>
         {expenses.map((expense) => {
@@ -41,13 +72,7 @@ export default function ExpensesSection({
             <Item.Root
               $hasAction
               key={expense.id}
-              onClick={() =>
-                onSelect?.({
-                  icon,
-                  title,
-                  type: 'expense',
-                })
-              }
+              onClick={() => handleOpenExpenseModal(expense)}
             >
               <Item.Box $align="center">
                 <Item.Icon height={28}>
