+++ front/src/renderer/src/views/pages/Dashboard/sections/CategoriesSection/index.tsx
@@ -1,7 +1,5 @@
 import { useDashboardCategoriesQuery } from '@renderer/app/hooks/queries/useDashboardCategoriesQuery';
 import Loader from '@renderer/views/components/Loader';
-import { useCallback, useState } from 'react';
-import CategoryModal from './CategoryModal';
 import ExpensesSection from './ExpensesSection';
 import IncomesSection from './IncomesSection';
 import { Container, Separator } from './styles';
@@ -13,20 +11,6 @@ export type SelectedCategory = {
 };
 
 export default function CategoriesSection() {
-  const [isOpenCategoryModal, setIsOpenCategoryModal] = useState(false);
-  const [selectedCategory, setSelectedCategory] =
-    useState<SelectedCategory | null>(null);
-
-  const handleOpenCategoryModal = useCallback((category: SelectedCategory) => {
-    setSelectedCategory(category);
-    setIsOpenCategoryModal(true);
-  }, []);
-
-  const handleCloseCategoryModal = useCallback(() => {
-    setSelectedCategory(null);
-    setIsOpenCategoryModal(false);
-  }, []);
-
   const {
     categories: { expenses, incomes },
     isLoading,
@@ -34,26 +18,14 @@ export default function CategoriesSection() {
 
   return (
     <Container>
-      {isOpenCategoryModal && (
-        <CategoryModal
-          open
-          onClose={handleCloseCategoryModal}
-          icon={selectedCategory!.icon}
-          title={selectedCategory!.title}
-          type={selectedCategory!.type}
-        />
-      )}
-
       {!isLoading && (
         <>
           <IncomesSection
             incomes={incomes}
-            onSelect={handleOpenCategoryModal}
           />
           <Separator />
           <ExpensesSection
             expenses={expenses}
-            onSelect={handleOpenCategoryModal}
           />
         </>
       )}
